[Minor] Composer package name should be `schumacherfm`
All other SchumacherFM Composer packages are in the `schumacherfm` namespace. However, this one is in `schumacher-fm`.

If this change is accepted, this repo will need to be resubmitted to Packagist. Also, the old package should be marked as "Abandoned" on Packagist.org so that people are redirected to the new name when trying to install using the old one.